pyproject.toml
[tool.poetry]
name = "bin_x"
version = "0.02"
description = ""
authors = ["K. D. Sunera Avinash Chandrasiri <kdsuneraavinash@gmail.com>",
    "Thumula Perera <thumula.17@cse.mrt.ac.lk>",
    "Anjala Dilhara <anjaladilhara@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
click = "^8.0.1"
pandas = "^1.3.1"
biopython = "^1.79"
numpy = "^1.21.1"
quadprog = "^0.1.8"
cvxopt = "^1.2.6"
scipy = "^1.7.1"
sklearn = "^0.0"
matplotlib = "^3.4.2"
seaborn = "^0.11.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.14.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
